test: consolidate grid event tests
Move cell-focus tests from keyboard-navigation.test.js into
events.test.js (renamed from activate-events.test.js) and rewrite
them to use real keyboard and mouse interactions.
